src/test/java/com/andrewauclair/todo/command/GroupCreateCommandTest.java
// Copyright (C) 2019 Andrew Auclair - All Rights Reserved
package com.andrewauclair.todo.command;

import org.jline.builtins.Completers;
import org.junit.jupiter.api.Test;

import java.util.Collections;
import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.jline.builtins.Completers.TreeCompleter.node;

class GroupCreateCommandTest extends CommandsBaseTestCase {
	@Test
	void verify_auto_complete_nodes() {
		GroupCreateCommand command = new GroupCreateCommand(tasks);

		List<Completers.TreeCompleter.Node> autoCompleteNodes = command.getAutoCompleteNodes();

		assertThat((Object) autoCompleteNodes).isEqualToComparingFieldByFieldRecursively(
				Collections.singletonList(node("mkgrp"))
		);
	}
}
